Move tests to evaluator folder.
